Update build scripts when only VS2022 is installed
This sets the default to check for VS2022 (v17) and if it's 16 or higher it uses the default paths starting with the vswhere path and adding the known path to MSBuild.exe. 

Tested on a machine with only VS2022 installed and on one with only VS2019 installed. 

Fixes the error when only VS2022 is installed.